requirements.txt
asgiref==3.4.1
Django==3.2.5
pytz==2021.1
sqlparse==0.4.1
django-extensions==3.2.3
pytest-django==4.7.0
pytest-cov==4.1.0
psycopg2-binary==2.9.9